requirements.txt
flask
flask-admin
flask-sqlalchemy
sqlalchemy
sqlalchemy-utils
passlib
arrow
colour
wtforms
python-dateutil
jinja2
